fibonacci numbers

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -5,14 +5,14 @@
  */
 int main(void)
 {
-int i, j, n;
+long int i, j, n;
 i = 1;
 j = 2;
 n = 0;
 while (n < 25)
 {
-printf("%d, ", i);
-printf("%d, ", j);
+printf("%ld, ", i);
+printf("%ld, ", j);
 i = i + j;
 j = i + j;
 n++;
